migrate list columns to render without expander cells
